fix(telegram-chatbot): cast \::bigint in subquery to avoid type deduction conflict
Postgres báo 'inconsistent types deduced for parameter \' vì:
- telegram_chatbot_settings.id_chatbot là INTEGER
- custom_chatbots.id là BIGINT
- \ vừa dùng cho VALUES (INTEGER) vừa cho WHERE cb.id = \ (BIGINT).
  Khi chatbotId=null (account-level toggle), Postgres không thể deduce
  kiểu nhất quán → 500 ở /api/ai/chatbot/telegram-account/chatbot/toggle.

Fix: ép \ thành bigint trong WHERE subquery. Subquery chỉ cần match
theo custom_chatbots.id nên cast bigint là phù hợp và rõ ràng.

+++ backend/src/repositories/chatbot/chatbotTelegram.repository.js
@@ -511,7 +511,7 @@ class ChatbotTelegramRepository {
        VALUES ($1, $2, $3, $4, $5,
                (SELECT NULLIF(BTRIM(cb.system_instruction), '')
                   FROM custom_chatbots cb
-                 WHERE cb.id = $2 AND cb.is_active = true))
+                 WHERE cb.id = $2::bigint AND cb.is_active = true))
        ON CONFLICT (id_telegram_account, id_chatbot) DO UPDATE SET
          is_enabled      = EXCLUDED.is_enabled,
          is_enabled_dm   = EXCLUDED.is_enabled_dm,
